requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
supabase==2.27.3
bcrypt==5.0.0
slowapi==0.1.9
solders>=0.21.0
solana>=0.34.0
anchorpy>=0.20.0
